Add files via upload
Copy and use the entire Search Registry folder to your local machine to keep the integrity of the original copy.


Run the Search-Registry.ps1 first to enable the search-registry function in PowerShell.


Search-Registry.ps1  is a registry search function that you can run in PowerShell. After running the script in the shell, type "help search-registry -full" to learn about the function.


Search-Registry-Add.ps1  is another script that calls the Search-Registry.ps1 to help search in registry by updating the the given variables in the script.